chore: rename & move offline-provider
