+++ services/offline/src/index.js
@@ -1,2 +1,2 @@
-export { OfflineProvider } from './components/OfflineProvider'
+export { OfflineProvider } from './lib/offline-provider'
 export { CacheableSection, useCacheableSection } from './lib/cacheable-section'
